// cosmowcl/src/app/shared/models/api/cosmoecm/richiediGenerazioneReportRequest.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * documento
 * API per la gestione dei documenti
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RichiediGenerazioneReportInputUtenteRequest } from './richiediGenerazioneReportInputUtenteRequest.d';


export interface RichiediGenerazioneReportRequest { 
    idPratica?: number;
    codiceTemplate?: string;
    codiceTipoDocumento?: string;
    mappaturaParametri?: string;
    nomeFile?: string;
    titolo?: string;
    autore?: string;
    formato?: string;
    sovrascrivi?: boolean;
    inputUtente?: Array<RichiediGenerazioneReportInputUtenteRequest>;
}